refactor: extract parsing module from add command
Extract from `cmd/add_cmd.rs` (1081 lines):
- `cmd/add_cmd/parsing.rs` (129 lines) — parse_date, parse_amount,
  calculate_balance

mod.rs retains Args, completer, interactive/quick modes, file ops,
and tests.

No behavioral changes.
